refactor: ui tests using latest powerapps bindings
