+++ tests/DevelopmentHub.Tests.Ui/Data/a solution merge awaiting manual merge activities.json
