+++ tests/DevelopmentHub.Tests.Ui/Develop/Create a solution.feature
@@ -4,23 +4,23 @@
 	I want to create solutions in Dynamics 365
 
 Scenario: Create a new solution
-	Given I am logged in to the Development Hub app as an admin
-	And I am viewing the Solutions sub area of the Develop area
-	When I select the New command
-	Then I can edit the following fields 
-	| Field                  |
-	| devhub_displayname        |
-	| devhub_uniquename         |
-	| devhub_stagingenvironment |
-	| devhub_description        |
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	When I open the sub area 'Solutions' under the 'Develop' area
+	And I select the 'New' command
+	Then I can edit the following fields
+		| Field                     |
+		| devhub_displayname        |
+		| devhub_uniquename         |
+		| devhub_stagingenvironment |
+		| devhub_description        |
 
 Scenario: Create a new issue without mandatory fields
-	Given I am logged in to the Development Hub app as an admin
-	And I am viewing the Solutions sub area of the Develop area
-	When I select the New command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	When I open the sub area 'Solutions' under the 'Develop' area
+	And I select the 'New' command
 	And I save the record
 	Then a mandatory field error is displayed on the following fields
-	| Field    |
-	| devhub_displayname        |
-	| devhub_uniquename         |
-	| devhub_stagingenvironment |
+		| Field                     |
+		| devhub_displayname        |
+		| devhub_uniquename         |
+		| devhub_stagingenvironment |
