+++ tests/DevelopmentHub.Tests.Ui/Develop/Create an environment.feature
@@ -4,20 +4,20 @@
 	I want to create environments in Dynamics 365
 
 Scenario: Create a new environment
-	Given I am logged in to the Development Hub app as an admin
-	And I am viewing the Environments sub area of the Develop area
-	When I select the New command
-	Then I can edit the following fields 
-	| Field    |
-	| devhub_name |
-	| devhub_url  |
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	When I open the sub area 'Environments' under the 'Develop' area
+	And I select the 'New' command
+	Then I can edit the following fields
+		| Field       |
+		| devhub_name |
+		| devhub_url  |
 
 Scenario: Create a new issue without mandatory fields
-	Given I am logged in to the Development Hub app as an admin
-	And I am viewing the Environments sub area of the Develop area
-	When I select the New command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	When I open the sub area 'Environments' under the 'Develop' area
+	And I select the 'New' command
 	And I save the record
 	Then a mandatory field error is displayed on the following fields
-	| Field    |
-	| devhub_name |
-	| devhub_url  |
+		| Field       |
+		| devhub_name |
+		| devhub_url  |
