+++ tests/DevelopmentHub.Tests.Ui/Develop/Reject a developed issue.feature
@@ -3,15 +3,9 @@
 	As a developer
 	I want to reject a solution merge
 
-Scenario: Open a solution merge awaiting review
-	Given I am logged in to the Development Hub app as an admin
-	And I have created an 'Awaiting Review' solution merge
-	And I have opened the solution merge
-	Then I can see the Reject command
-
 Scenario: Approve a solution merge awaiting review
-	Given I am logged in to the Development Hub app as an admin
-	And I have created an 'Awaiting Review' solution merge
-	And I have opened the solution merge
-	When I select the Reject command
-	Then I can't see the Reject command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	And I have created 'a solution merge awaiting review'
+	And I have opened 'the solution merge'
+	When I select the 'Reject' command
+	Then I can not see the 'Reject' command
