+++ tests/DevelopmentHub.Tests.Ui/Develop/Start developing an issue.feature
@@ -4,19 +4,19 @@
 	I want to start developing an issue that has been created in Dynamics 365
 
 Scenario: An unsaved issue
-	Given I am logged in to the Development Hub app as an admin
-	And I am viewing the Issues sub area of the Issues area
-	When I select the New command
-	Then I can't see the Start Developing command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	When I open the sub area 'Issues' under the 'Issues' area
+	And I select the 'New' command
+	Then I can not see the 'Develop' command
 
 Scenario: A 'to do' issue
-	Given I am logged in to the Development Hub app as an admin
-	And I have created a 'To Do' issue
-	And I have opened the issue
-	Then I can see the Start Developing command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	And I have created 'a new issue'
+	And I have opened 'the issue'
+	Then I can see the 'Develop' command
 
 Scenario: An issue that is not 'to do'
-	Given I am logged in to the Development Hub app as an admin
-	And I have created a 'In Progress' issue
-	And I have opened the issue
-	Then I can't see the Start Developing command
+	Given I am logged in to the 'Development Hub' app as 'an admin'
+	And I have created 'an in-progress issue'
+	And I have opened 'the issue'
+	Then I can not see the 'Develop' command
