+++ tests/DevelopmentHub.Tests.Ui/Properties/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
-using Xunit;
+using NUnit.Framework;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
@@ -13,7 +13,7 @@ using Xunit;
 [assembly: AssemblyCopyright("Copyright © Development Hub 2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
-[assembly: AssemblyTrait("TestCategory", "UI")]
+[assembly: Category("UI")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from
