+++ tests/DevelopmentHub.Tests.Ui/packages.config
@@ -1,49 +1,44 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="BoDi" version="1.4.1" targetFramework="net462" />
-  <package id="Capgemini.Test.Xrm" version="0.0.7" targetFramework="net462" />
-  <package id="Capgemini.Test.Xrm.Uci" version="0.0.9" targetFramework="net462" />
+  <package id="Capgemini.PowerApps.SpecFlowBindings" version="0.9.12" targetFramework="net462" />
   <package id="Cucumber.Messages" version="6.0.1" targetFramework="net462" />
-  <package id="Dynamics365.UIAutomation.Api" version="9.1.0.12013" targetFramework="net462" />
+  <package id="Dynamics365.UIAutomation.Api" version="9.1.0.23435" targetFramework="net462" />
+  <package id="FluentAssertions" version="5.10.3" targetFramework="net462" />
   <package id="Gherkin" version="6.0.0" targetFramework="net462" />
   <package id="Google.Protobuf" version="3.7.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.2" targetFramework="net462" developmentDependency="true" />
-  <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.2" targetFramework="net462" developmentDependency="true" />
-  <package id="Microsoft.CodeQuality.Analyzers" version="2.9.2" targetFramework="net462" developmentDependency="true" />
-  <package id="Microsoft.NetCore.Analyzers" version="2.9.2" targetFramework="net462" developmentDependency="true" />
-  <package id="Microsoft.NetFramework.Analyzers" version="2.9.2" targetFramework="net462" developmentDependency="true" />
+  <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="3.3.0" targetFramework="net462" developmentDependency="true" />
+  <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="3.3.0" targetFramework="net462" developmentDependency="true" />
+  <package id="Microsoft.CodeQuality.Analyzers" version="3.3.0" targetFramework="net462" developmentDependency="true" />
+  <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.26" targetFramework="net462" />
+  <package id="Microsoft.NetCore.Analyzers" version="3.3.0" targetFramework="net462" developmentDependency="true" />
+  <package id="Microsoft.NetFramework.Analyzers" version="3.3.0" targetFramework="net462" developmentDependency="true" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net462" />
-  <package id="NUnit" version="3.10.1" targetFramework="net462" />
-  <package id="Selenium.Chrome.WebDriver" version="2.45" targetFramework="net462" />
+  <package id="NUnit" version="3.12.0" targetFramework="net462" />
+  <package id="NUnit3TestAdapter" version="3.17.0" targetFramework="net462" />
+  <package id="Otp.NET" version="1.2.1" targetFramework="net462" />
+  <package id="Selenium.Chrome.WebDriver" version="85.0.0" targetFramework="net462" />
   <package id="Selenium.Support" version="3.141.0" targetFramework="net462" />
   <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net462" />
-  <package id="SpecFlow" version="3.1.82" targetFramework="net462" />
-  <package id="SpecFlow.Tools.MsBuild.Generation" version="3.1.82" targetFramework="net462" />
-  <package id="SpecFlow.xUnit" version="3.1.82" targetFramework="net462" />
+  <package id="SpecFlow" version="3.4.14" targetFramework="net462" />
+  <package id="SpecFlow.NUnit" version="3.4.14" targetFramework="net462" />
+  <package id="SpecFlow.Tools.MsBuild.Generation" version="3.4.14" targetFramework="net462" />
   <package id="SpecRun.Runner" version="1.8.5" targetFramework="net462" />
   <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net462" developmentDependency="true" />
+  <package id="System.IO" version="4.3.0" targetFramework="net462" />
   <package id="System.Net.Http" version="4.3.4" targetFramework="net462" />
-  <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net462" />
-  <package id="System.Reflection.Emit.Lightweight" version="4.3.0" targetFramework="net462" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net462" />
+  <package id="System.Reflection.Emit" version="4.7.0" targetFramework="net462" />
+  <package id="System.Reflection.Emit.Lightweight" version="4.7.0" targetFramework="net462" />
+  <package id="System.Runtime" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net462" />
   <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net462" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.3.1" targetFramework="net462" />
   <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net462" />
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net462" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.3" targetFramework="net462" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net462" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net462" />
   <package id="System.ValueTuple" version="4.5.0" targetFramework="net462" />
-  <package id="Text.Analyzers" version="2.6.3" targetFramework="net462" developmentDependency="true" />
+  <package id="Text.Analyzers" version="2.6.4" targetFramework="net462" developmentDependency="true" />
   <package id="Utf8Json" version="1.3.7" targetFramework="net462" />
-  <package id="Validation" version="2.4.18" targetFramework="net462" />
-  <package id="xunit" version="2.4.1" targetFramework="net462" />
-  <package id="xunit.abstractions" version="2.0.3" targetFramework="net462" />
-  <package id="xunit.analyzers" version="0.10.0" targetFramework="net462" />
-  <package id="xunit.assert" version="2.4.1" targetFramework="net462" />
-  <package id="xunit.core" version="2.4.1" targetFramework="net462" />
-  <package id="xunit.extensibility.core" version="2.4.1" targetFramework="net462" />
-  <package id="xunit.extensibility.execution" version="2.4.1" targetFramework="net462" />
-  <package id="xunit.runner.visualstudio" version="2.4.1" targetFramework="net462" developmentDependency="true" />
-  <package id="Xunit.SkippableFact" version="1.3.12" targetFramework="net462" />
-  <package id="YamlDotNet" version="6.0.0" targetFramework="net462" />
+  <package id="YamlDotNet" version="8.1.2" targetFramework="net462" />
 </packages>
