--- tests/DevelopmentHub.Tests.Ui/spec-xrm.yaml
@@ -1,10 +1,0 @@
-url: "https://devhubci.crm11.dynamics.com"
-browser: Chrome
-users: 
-  - username: "service@devhubci.onmicrosoft.com"
-    password: ""
-    isAdfs: false
-    alias: "an admin"
-apps: 
-  - name: "Development Hub"
-    id: "438304b1-8879-e911-a9b6-002248073e1b"
