+++ tests/DevelopmentHub.Tests.Ui/specflow.json
@@ -1,5 +1,5 @@
 {
   "stepAssemblies": [
-    { "assembly": "Capgemini.Test.Xrm.Uci" }
+    { "assembly": "Capgemini.PowerApps.SpecFlowBindings" }
   ]
 }
