--- tests/DevelopmentHub.Tests.Ui/xunit.runner.json
@@ -1,3 +1,0 @@
-{
-  "shadowCopy": false
-}
